refactor(beizer-codemod): include stack transform function to refactor
