+++ packages/bezier-codemod/src/App.tsx
@@ -21,7 +21,7 @@ import project from './project.js'
 import enumMemberToStringLiteral from './transforms/enum-member-to-string-literal/transform.js'
 import iconNameToBezierIcon from './transforms/icon-name-to-bezier-icons/transform.js'
 import iconsToBezierIcons from './transforms/icons-to-bezier-icons/transform.js'
-import removeAlphaFromAlphaStack from './transforms/remove-alpha-from-alpha-stack.js'
+import removeAlphaFromAlphaStack from './transforms/v2-remove-alpha-from-alpha-stack/transform.js'
 import foundationToCssVariableBorder from './transforms/v2-foundation-to-css-variable/border.js'
 import foundationToCssVariableElevation from './transforms/v2-foundation-to-css-variable/elevation.js'
 import foundationToCssVariableRounding from './transforms/v2-foundation-to-css-variable/rounding.js'
