+++ packages/bezier-codemod/src/transforms/v2-remove-alpha-from-alpha-stack/fixtures/stack2.output.tsx
