+++ packages/bezier-codemod/src/transforms/v2-remove-alpha-from-alpha-stack/fixtures/stack4.input.tsx
