+++ packages/bezier-codemod/src/transforms/v2-remove-alpha-from-alpha-stack/transform.ts
@@ -3,7 +3,7 @@ import {
   SyntaxKind,
 } from 'ts-morph'
 
-import { renameNamedImport } from '../utils/import.js'
+import { renameNamedImport } from '../../utils/import.js'
 
 const LEGACY_STACKS = ['Stack', 'HStack', 'VStack', 'StackItem', 'Spacer']
 const ALPHA_STACKS = ['AlphaStack']
